diverse Änderungen

+++ plugins/server/pages/dsgvo.server-edit.php
@@ -3,12 +3,14 @@ echo rex_view::title($this->i18n('dsgvo'));
 
 	$func = rex_request('func', 'string');
 	$domain = rex_request('domain', 'string');
+	$result = rex_request('result', 'string', false);
 
-	if ($func == '' || $func == "domain_delete") { 
+	if (($func == '' && !$result) || $func == "domain_delete") { 
 
 		if($func == 'domain_delete') {
     		$oid = rex_request('oid', 'int');
 			$delete = rex_sql::factory()->setQuery('DELETE FROM rex_dsgvo_server_project WHERE id = :oid',array(':oid' => $oid));
+			$delete = rex_sql::factory()->setDebug(1)->setQuery('DELETE FROM rex_dsgvo_server WHERE domain = :domain',array(':domain' => $domain));
 			echo rex_view::success( $this->i18n('dsgvo_server_domain_deleted'));
     	}	
 
@@ -17,7 +19,7 @@ echo rex_view::title($this->i18n('dsgvo'));
 		$list->addTableAttribute('class', 'table-striped');
 		$list->setNoRowsMessage($this->i18n('dsgvo_server_norows_message'));
 		
-		// icon column
+		// icon column (Domain hinzufügen bzw. bearbeiten)
 		$thIcon = '<a href="'.$list->getUrl(['func' => 'domain_add']).'"><i class="rex-icon rex-icon-add-action"></i></a>';
 		$tdIcon = '<i class="rex-icon fa-file-text-o"></i>';
 		$list->addColumn($thIcon, $tdIcon, 0, ['<th class="rex-table-icon">###VALUE###</th>', '<td class="rex-table-icon">###VALUE###</td>']);
@@ -25,15 +27,11 @@ echo rex_view::title($this->i18n('dsgvo'));
 		
 		$list->setColumnLabel('domain', $this->i18n('dsgvo_server_domain_column_domain'));
 		$list->setColumnParams('domain', ['id' => '###id###', 'func' => 'domain_edit']);
-
-		$list->setColumnLabel('api_key', $this->i18n('dsgvo_server_domain_column_api_key'));	
-		
-		$list->removeColumn('id');
-		$list->removeColumn('updatedate');
-
-		
+			
 		$list->addColumn($this->i18n('dsgvo_server_domain_column_manage_text'), $this->i18n('dsgvo_server_domain_column_manage_text'), 3);
 		$list->setColumnParams($this->i18n('dsgvo_server_domain_column_manage_text'), ['data_id' => '###id###', 'func' => 'domain_details', 'domain' => '###domain###']);
+
+		$list->setColumnLabel('api_key', $this->i18n('dsgvo_server_domain_column_api_key'));
 
 		$list->addColumn('last_call', '');
 		$list->setColumnLabel('last_call', $this->i18n('dsgvo_server_domain_column_last_call'));
@@ -47,19 +45,177 @@ echo rex_view::title($this->i18n('dsgvo'));
 		});
 		
 		$list->addColumn('domain_delete', '<i class="rex-icon rex-icon-delete"></i> ' . $this->i18n('dsgvo_server_domain_column_delete'), -1, ['', '<td class="rex-table-action">###VALUE###</td>']);
-    	$list->setColumnParams('domain_delete', ['func' => 'domain_delete', 'oid' => '###id###']);
+    	$list->setColumnParams('domain_delete', ['func' => 'domain_delete', 'oid' => '###id###', 'domain' => '###domain###']);
     	$list->addLinkAttribute('domain_delete', 'data-confirm', $this->i18n('dsgvo_server_domain_delete_confirm'));
+
+    	$list->removeColumn('id');
+		$list->removeColumn('updatedate');
 		
 		$content1 = $list->get();
 		
 		$fragment = new rex_fragment();
 		$fragment->setVar('class', "info", false);
-		$fragment->setVar('title', $this->i18n('dsgvo_client_project_server_title'), false);
+		$fragment->setVar('title', $this->i18n('dsgvo_server_domain_list_title'), false);
 		$fragment->setVar('content', $content1, false);
 		$content1 = $fragment->parse('core/page/section.php');
 		
 		echo $content1;
 		// Domain-Übersicht ENDE //
+
+	} else if ($func == 'domain_add' || $func == 'domain_edit') { 
+		
+		// Domain bearbeiten ANFANG //
+
+		$id = rex_request('id', 'int');
+		
+		if ($func == 'domain_edit') {
+			$formLabel = $this->i18n('dsgvo_server_text_edit');
+		} elseif ($func == 'domain_add') {
+			$formLabel = $this->i18n('dsgvo_server_text_add');
+		}
+		
+		$form = rex_form::factory(rex::getTablePrefix().'dsgvo_server_project', '', 'id='.$id);
+		
+		//Start - add domain-field
+		$field = $form->addTextField('domain');
+		$field->setLabel($this->i18n('dsgvo_server_domain_column_domain'));
+		$field->setNotice($this->i18n('dsgvo_server_domain_column_domain_note'));
+		//End - add domain-field
+
+		//Start - add domain-field
+		$field = $form->addTextField('api_key');
+		$field->setLabel($this->i18n('dsgvo_server_domain_column_api_key'));
+		$field->setNotice($this->i18n('dsgvo_server_domain_column_api_key_note', md5(time())));
+		//End - add domain-field
+		
+		if ($func == 'domain_edit') {
+			$form->addParam('id', $id);
+		}
+
+		$content3 = $form->get();
+
+		$fragment = new rex_fragment();
+		$fragment->setVar('class', 'edit', false);
+		$fragment->setVar('title', $formLabel, false);
+		$fragment->setVar('body', $content3, false);
+		$content3 = $fragment->parse('core/page/section.php');
+
+		echo $content3;
+		// Domain bearbeiten ENDE //
+
+	} else if ($func == 'domain_details' || $func == 'text_copy_default' || $func == 'text_delete' || $func == 'set_text_status' || ($func == '' && $result)) {
+
+		if($func == 'text_delete') {
+    		$oid = rex_request('oid', 'int');
+			$delete = rex_sql::factory()->setQuery('DELETE FROM rex_dsgvo_server WHERE id = :oid',array(':oid' => $oid));
+			echo rex_view::success( $this->i18n('dsgvo_server_text_deleted'));
+		}	
+		if($func == 'text_copy_default') {
+			$query = 'INSERT INTO rex_dsgvo_server (`domain`, `lang`, `name`, `keyword`, `text`, `source`, `source_url`, `prio`, `status`, `updatedate`) (SELECT :domain AS `domain`, `lang`, `name`, `keyword`, `text`, `source`, `source_url`, `prio`, 0 as `status`, NOW() FROM rex_dsgvo_server WHERE domain = "default")';
+			rex_sql::factory()->setDebug(0)->setQuery($query, [":domain" => $domain]);
+			echo rex_view::success( $this->i18n('dsgvo_server_text_default_copied'));
+		}	
+				
+		// Offline / Online schalten
+		if ($func == 'set_text_status') {
+			$status = (rex_request('oldstatus', 'int') + 1) % 2;
+			$msg = $status == 1 ? 'dsgvo_server_status_activate' : 'dsgvo_server_status_deactivate';
+			$oid = rex_request("oid", "int");
+			$update = rex_sql::factory()->setQuery('UPDATE rex_dsgvo_server SET status = :status WHERE id = :oid',array(':status' => $status, ':oid' => $oid))->execute(array(':status' => $status, ':oid' => $oid));
+			if ($update) {
+				echo rex_view::success($this->i18n($msg . '_success', $name));
+			} else {
+				echo rex_view::error($this->i18n($msg . '_error', $name));
+			}
+		}
+		
+
+		$list = rex_list::factory('SELECT * FROM `'.rex::getTablePrefix().'dsgvo_server` WHERE domain = "'.$domain.'" ORDER BY `prio` ASC');
+		$list->addTableAttribute('class', 'table-striped');
+		$list->setNoRowsMessage($this->i18n('dsgvo_server_norows_message'));
+		
+		// icon column
+		$thIcon = '<a href="'.$list->getUrl(['func' => 'text_add', 'domain' => $domain]).'"><i class="rex-icon rex-icon-add-action"></i></a>';
+		$tdIcon = '<i class="rex-icon fa-file-text-o"></i>';
+		$list->addColumn($thIcon, $tdIcon, 0, ['<th class="rex-table-icon">###VALUE###</th>', '<td class="rex-table-icon">###VALUE###</td>']);
+		$list->setColumnParams($thIcon, ['func' => 'text_edit', 'domain' => $domain, 'id' => '###id###']);
+		
+		//$list->setColumnLabel('name', $this->i18n('sets_column_name'));
+		$list->setColumnLabel('type', $this->i18n('sets_column_type'));		
+		$list->setColumnParams('name', ['id' => '###id###', 'func' => 'text_edit']);
+
+		$list->setColumnLabel('domain', $this->i18n('dsgvo_server_text_column_domain'));
+		
+		$list->setColumnLabel('lang', $this->i18n('dsgvo_server_text_column_lang'));
+		$list->setColumnLabel('name', $this->i18n('dsgvo_server_text_column_name'));
+		$list->setColumnLabel('source', $this->i18n('dsgvo_server_text_column_source'));
+		$list->setColumnLabel('prio', $this->i18n('dsgvo_server_text_column_prio'));
+		$list->setColumnLabel('status', $this->i18n('dsgvo_server_text_column_status'));
+		$list->setColumnParams('status', ['func' => 'set_text_status', 'oldstatus' => '###status###', 'oid' => '###id###', 'domain' => '###domain###']);
+		$list->setColumnFormat('status', 'custom', function ($params) {
+			$list = $params['list'];  
+			if ($params['value'] == "") {
+				$str = rex_i18n::msg('dsgvo_client_text_column_status_invalid');
+			} elseif ($params['value'] == 1) {
+				$str = $list->getColumnLink('status', '<span class="rex-online"><i class="rex-icon rex-icon-online"></i> ' . rex_i18n::msg('dsgvo_client_text_column_status_is_online') . '</span>');
+			} else {
+				$str = $list->getColumnLink('status', '<span class="rex-offline"><i class="rex-icon rex-icon-offline"></i> ' . rex_i18n::msg('dsgvo_client_text_column_status_is_offline') . '</span>');
+			}
+			return $str;
+		});
+		$list->setColumnLabel('updatedate', $this->i18n('dsgvo_server_text_column_updatedate'));
+
+		$list->addColumn('text_delete', '<i class="rex-icon rex-icon-delete"></i> ' . $this->i18n('dsgvo_server_text_column_delete'), -1, ['', '<td class="rex-table-action">###VALUE###</td>']);
+    	$list->setColumnParams('text_delete', ['func' => 'text_delete', 'oid' => '###id###', 'domain' => '###domain###']);
+    	$list->addLinkAttribute('text_delete', 'data-confirm', $this->i18n('dsgvo_server_text_delete_confirm'));
+		
+		$list->removeColumn('keyword');
+		$list->removeColumn('id');
+		$list->removeColumn('text');
+		$list->removeColumn('custom_text');
+		$list->removeColumn('source_url');
+		$list->removeColumn('code');
+		
+		$content4 = $list->get();
+		
+		$fragment = new rex_fragment();
+		$fragment->setVar('class', "info", false);
+		$fragment->setVar('title', $this->i18n('dsgvo_server_text_title'), false);
+		$fragment->setVar('content', $content4, false);
+		$content4 = $fragment->parse('core/page/section.php');
+		
+		echo $content4;
+
+		// LOGS
+		$domain = rex_request('domain', 'string', "");
+		$list = rex_list::factory('SELECT * FROM rex_dsgvo_server_log WHERE domain = "'.$domain.'" ORDER BY createdate DESC LIMIT 30', 10, "domain");
+
+		$list->setColumnLabel('id', $this->i18n('dsgvo_server_project_log_id'));
+		$list->setColumnLabel('domain', $this->i18n('dsgvo_server_project_log_domain'));
+		$list->setColumnLabel('status', $this->i18n('dsgvo_server_project_log_status'));
+		$list->setColumnLabel('createdate', $this->i18n('dsgvo_server_project_log_createdate'));
+		$list->removeColumn('raw');
+		
+		$fragment = new rex_fragment();
+		$fragment->setVar('class', 'default', false);
+		$fragment->setVar('title', $this->i18n('dsgvo_server_project_log_title'), false);
+		$fragment->setVar('body', $list->get(), false);
+		$content5 = $fragment->parse('core/page/section.php');
+		echo $content5;
+		// LOGS END
+
+
+		if(rex_request('domain','string') != 'default') {
+			$buttons = '<a class="btn btn-edit" href="index.php?page=dsgvo/server-edit&func=text_copy_default&domain='.$domain.'">' . rex_i18n::msg('dsgvo_server_default_text_copy') . '</a>';
+
+			$fragment = new rex_fragment();
+			$fragment->setVar('class', 'default', false);
+			$fragment->setVar('title', $this->i18n('dsgvo_server_default_title'), false);
+			$fragment->setVar('body', $buttons, false);
+			// $fragment->setVar('buttons', , false);
+			echo $fragment->parse('core/page/section.php');
+		}
+
 
 	} else if ($func == 'text_edit' || $func == 'text_add') { // Wenn von einer Domain Texte verwaltet werden
 				
@@ -153,7 +309,10 @@ echo rex_view::title($this->i18n('dsgvo'));
 		if ($func == 'text_edit') {
 			$form->addParam('id', $id);
 		}
-		
+
+		$form->addParam("domain", $domain);
+		$form->addParam("result", "text_added");
+
 		$content2 = $form->get();
 
 		$fragment = new rex_fragment();
@@ -165,127 +324,5 @@ echo rex_view::title($this->i18n('dsgvo'));
 		echo $content2;
 		// Text bearbeiten ENDE //
 
-	} else if ($func == 'domain_add' || $func == 'domain_edit') { 
-		
-		// Domain bearbeiten ANFANG //
-
-		$id = rex_request('id', 'int');
-		
-		if ($func == 'domain_edit') {
-			$formLabel = $this->i18n('dsgvo_server_text_edit');
-		} elseif ($func == 'domain_add') {
-			$formLabel = $this->i18n('dsgvo_server_text_add');
-		}
-		
-		$form = rex_form::factory(rex::getTablePrefix().'dsgvo_server_project', '', 'id='.$id);
-		
-		//Start - add domain-field
-		$field = $form->addTextField('domain');
-		$field->setLabel($this->i18n('dsgvo_server_domain_column_domain'));
-		$field->setNotice($this->i18n('dsgvo_server_domain_column_domain_note'));
-		//End - add domain-field
-
-		//Start - add domain-field
-		$field = $form->addTextField('api_key');
-		$field->setLabel($this->i18n('dsgvo_server_domain_column_api_key'));
-		$field->setNotice($this->i18n('dsgvo_server_domain_column_api_key_note', md5(time())));
-		//End - add domain-field
-		
-		if ($func == 'domain_edit') {
-			$form->addParam('id', $id);
-		}
-
-		$content3 = $form->get();
-
-		$fragment = new rex_fragment();
-		$fragment->setVar('class', 'edit', false);
-		$fragment->setVar('title', $formLabel, false);
-		$fragment->setVar('body', $content3, false);
-		$content3 = $fragment->parse('core/page/section.php');
-
-		echo $content3;
-		// Domain bearbeiten ENDE //
-
-	} else if ($func == 'domain_details') { 
-		
-		// Offline / Online schalten
-		if ($func == 'setstatus') {
-			$status = (rex_request('oldstatus', 'int') + 1) % 2;
-			$msg = $status == 1 ? 'dsgvo_server_status_activate' : 'dsgvo_server_status_deactivate';
-			$oid = rex_request("oid", "int");
-			$update = rex_sql::factory()->setQuery('UPDATE rex_dsgvo_server SET status = :status WHERE id = :oid',array(':status' => $status, ':oid' => $oid))->execute(array(':status' => $status, ':oid' => $oid));
-			if ($update) {
-				echo rex_view::success($this->i18n($msg . '_success', $name));
-			} else {
-				echo rex_view::error($this->i18n($msg . '_error', $name));
-			}
-			$func = '';
-		}
-
-		$list = rex_list::factory('SELECT * FROM `'.rex::getTablePrefix().'dsgvo_server` WHERE domain = "'.$domain.'" ORDER BY `prio` ASC');
-		$list->addTableAttribute('class', 'table-striped');
-		$list->setNoRowsMessage($this->i18n('dsgvo_server_norows_message'));
-		
-		// icon column
-		$thIcon = '<a href="'.$list->getUrl(['func' => 'text_add', 'domain' => $domain]).'"><i class="rex-icon rex-icon-add-action"></i></a>';
-		$tdIcon = '<i class="rex-icon fa-file-text-o"></i>';
-		$list->addColumn($thIcon, $tdIcon, 0, ['<th class="rex-table-icon">###VALUE###</th>', '<td class="rex-table-icon">###VALUE###</td>']);
-		$list->setColumnParams($thIcon, ['func' => 'text_edit', 'domain' => $domain, 'id' => '###id###']);
-		
-		//$list->setColumnLabel('name', $this->i18n('sets_column_name'));
-		$list->setColumnLabel('type', $this->i18n('sets_column_type'));		
-		$list->setColumnParams('name', ['id' => '###id###', 'func' => 'text_edit']);
-
-		$list->setColumnLabel('domain', $this->i18n('dsgvo_server_text_column_domain'));
-		
-		$list->setColumnLabel('lang', $this->i18n('dsgvo_server_text_column_lang'));
-		$list->setColumnLabel('name', $this->i18n('dsgvo_server_text_column_name'));
-		$list->setColumnLabel('source', $this->i18n('dsgvo_server_text_column_source'));
-		$list->setColumnLabel('prio', $this->i18n('dsgvo_server_text_column_prio'));
-		$list->setColumnLabel('status', $this->i18n('dsgvo_server_text_column_status'));
-		$list->setColumnParams('status', ['func' => 'setstatus', 'oldstatus' => '###status###', 'oid' => '###id###']);
-		$list->setColumnFormat('status', 'custom', function ($params) {
-			$list = $params['list'];  
-			if ($params['value'] == "") {
-				$str = rex_i18n::msg('cronjob_status_invalid');
-			} elseif ($params['value'] == 1) {
-				$str = $list->getColumnLink('status', '<span class="rex-online"><i class="rex-icon rex-icon-online"></i> ' . rex_i18n::msg('dsgvo_client_text_column_status_is_online') . '</span>');
-			} else {
-				$str = $list->getColumnLink('status', '<span class="rex-offline"><i class="rex-icon rex-icon-offline"></i> ' . rex_i18n::msg('dsgvo_client_text_column_status_is_offline') . '</span>');
-			}
-			return $str;
-		});
-		$list->setColumnLabel('updatedate', $this->i18n('dsgvo_server_text_column_updatedate'));
-		
-		$list->removeColumn('keyword');
-		$list->removeColumn('id');
-		$list->removeColumn('text');
-		$list->removeColumn('custom_text');
-		$list->removeColumn('source_url');
-		$list->removeColumn('code');
-		
-		$content4 = $list->get();
-		
-		$fragment = new rex_fragment();
-		$fragment->setVar('class', "info", false);
-		$fragment->setVar('title', $this->i18n('dsgvo_server_text_title'), false);
-		$fragment->setVar('content', $content4, false);
-		$content4 = $fragment->parse('core/page/section.php');
-		
-		echo $content4;
-
-		// LOGS
-		$domain = rex_request('domain', 'string', "");
-		$list = rex_list::factory('SELECT * FROM rex_dsgvo_server_log WHERE domain = "'.$domain.'" ORDER BY createdate DESC LIMIT 30', 10, "domain");
-		
-		$fragment = new rex_fragment();
-		$fragment->setVar('class', 'default', false);
-		$fragment->setVar('title', $this->i18n('dsgvo_server_project_log_title'), false);
-		$fragment->setVar('body', $list->get(), false);
-		$content5 = $fragment->parse('core/page/section.php');
-
-		echo $content5;
-		// LOGS END
-
-	}
+	}  
 ?>
